+++ Examples/Install-Module.ps1
@@ -11,7 +11,7 @@ try
     . "$PSScriptRoot\Cmdlets\Get-ISHBootstrapperContextValue.ps1"
     . "$cmdletsPaths\Helpers\Invoke-CommandWrap.ps1"
 
-    $computerName=Get-ISHBootstrapperContextValue -ValuePath "ComputerName"
+    $computerName=Get-ISHBootstrapperContextValue -ValuePath "ComputerName" -DefaultValue $null
 
     if(-not $computerName)
     {